% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGenomicData.R
\name{getAncestryPCs}
\alias{getAncestryPCs}
\title{Retrieve ancestry PCs for a set of persons}
\usage{
getAncestryPCs(cdm, person_ids = NULL, n_pcs = 10L)
}
\arguments{
\item{cdm}{A CDMConnector cdm_reference object.}

\item{person_ids}{Integer vector. If NULL, all persons returned.}

\item{n_pcs}{Number of PCs to return (1-10). Default 10.}
}
\value{
A data.frame with columns: person_id, pc1, ..., pc<n_pcs>
}
\description{
Retrieve ancestry PCs for a set of persons
}
