% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecate_client.R
\name{hecate_get_concept}
\alias{hecate_get_concept}
\title{Get a concept by ID from Hecate}
\usage{
hecate_get_concept(concept_id, client = NULL)
}
\arguments{
\item{concept_id}{Integer OMOP concept_id.}

\item{client}{A \code{hecate_client} object; created from env vars if NULL.}
}
\value{
Parsed concept details (list).
}
\description{
Get a concept by ID from Hecate
}
