% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecate_client.R
\name{hecate_get_relationships}
\alias{hecate_get_relationships}
\title{Get concept relationships from Hecate}
\usage{
hecate_get_relationships(concept_id, client = NULL)
}
\arguments{
\item{concept_id}{Integer OMOP concept_id.}

\item{client}{A \code{hecate_client} object; created from env vars if NULL.}
}
\value{
Parsed relationship list.
}
\description{
Get concept relationships from Hecate
}
