% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc_utils.R
\name{is_ndc_like}
\alias{is_ndc_like}
\title{Test whether a string looks like an NDC code}
\usage{
is_ndc_like(x)
}
\arguments{
\item{x}{Character string to test.}
}
\value{
Logical.
}
\description{
Heuristic: after stripping hyphens, spaces, and asterisks, is it 9-11 digits?
Also matches patterns like "59630*70248" or "00944-2620-01".
}
