% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_runner.R
\name{split_sql_statements}
\alias{split_sql_statements}
\title{Split SQL script into single statements (semicolon-terminated, ignore in strings)}
\usage{
split_sql_statements(sql)
}
\description{
Split SQL script into single statements (semicolon-terminated, ignore in strings)
}
